This PR does three things:

a. Fixes VectorMontonixer having too few polygon points to ensure that small inner burns are not mistakenly considered outside.
b. Fixes Rasters not being considered for inner burns, and treats rasters correctly as being inside even if it extends outside;
c. Improves performance by maintaining burns_done at a group level rather than iterating over segments to find out.