@Sophist-UK 2022-01-24
I have tested this with rasters and images and dots.

VectorMontonizer is still using brute force approach (with 1000 polygon points rather than 100). We should consider using a more intelligent choice for:

a. polygon points - based on path curvature and length. Long straight lines don't need intermediate points, long curves will need intermediate points based on curvature, short path segments will need to be approximated by straight lines, and then points added based on changes of direction and / or maximum deviation (like 3d printer slicers do to reduce stuttering).

b. inner points - based on absolute size or size relative to the outer path.