@Sophist-UK 2022-01-24
Performance test results (for a leather pattern with one outer and a couple of hundred inners):
|      | Main | This PR |
| --- | --- | --- |
| Inner | 4.051s | 1.970s |
| Greedy | 3.821s | 0.424s |