@Sophist-UK 2022-01-24
Need to fix a poorly chosen algorithm in CutGroup burns_done.