@Sophist-UK 2022-01-24
P.S. I think this is a good pre-requisite for my planned enhancement to break dispersed rasters into several smaller images.