@tatarize 2022-01-24
We may at some point need to consider the utilization of the plotplanner to cut up the paths. It's actually not that much slower from some of the checks against this with vpype and actually decided very definitional with regards to the shape. This might actually need an implementation of the sort of phase 2 potrace algorithm which would be rather helpful with some of these conversion.

http://potrace.sourceforge.net/potrace.pdf gives the Sellinger used for the straight path line detection metric. And give that we can actually get point by point curve sections generated from the zingl-bresenham line draw algorithm we could utilize a version of this to result in more-or-less straight line sections to approximate the curves we are using. This would be a weird method of linearizing but rather than a sample rate, we'd actually get remarkably linear sections that depend only and entirely on the native resolution at which we are applying the curves. So the zingl-bresenham curve drawing algorithms would give us a series of single step ortho and diagonal steps, and we'd combine these into line sections.

With a clear representation of the shape in linear segments at native resolution we'd have a shape that given your resolution would be identical to the original shape in perfectly sized linear segments. For some of the smarter boards this would be extremely helpful as well, as we could get nearly perfect curves made out of linear segments. Likewise for the vector monotonized this would also work to help provide solid insideness definitions.  Since the only time we'd overlap these points is if they actively intersect. It would give us perfect sets of lines at exactly the resolution we needed.  Since it would derive from the curve drawing algorithm and line plotting exactly from that plotted path.